Made a pretty big change to the grammar.
Turns out, the current grammar was really hard to parse, & changing
it around made it left recursive.

I've fixed this by GREATLY simplifying the expression syntax, which actually
makes the grammar a lot more powerful. The only issue now is that type errors won't
be caught in the parser.

This means, for example, we can now perform operations like
if (1) {...}, which is obviously invalid.

On the plus side, this simplifies the parser a great deal.